Adding command line argument for changing config dir.
Unfortunately the Java samples are method-specific where each method has its own main(), so there's a lot more noise here than I'd like. The important changes are in src/shopping/common and in the roots of the API-specific directories. The rest are a combination of lifting the args to BaseSample's constructor and results of the autoformatter.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=155077289