Retrieve all appropriate information from the Content API.
This change moves the website URL from being part of the configuration to being
retrieved from the Content API.

In addition, if no Merchant Center ID is provided in the configuration, then we
will default to the first Merchant Center ID to which the currently
authenticated user has access. With this, the configuration file is completely
optional, and all that is needed is authentication credentials.

Since we can supply authentication credentials via Google Application Default
Credentials, add a flag that allows for the samples to be run without a
configuration directory at all.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157070310