Extend these demos to handle MCA accounts.
The old split demos for the Accounts service now have one entry point that
first runs the primaryAccount demo on any given account, and if that account is
an MCA, it then runs the MCA portion.

The Accountstatuses service demo is updated similarly where there's a portion
that runs on all accounts, and then an MCA-only portion. The MCA-only portion
is new to this change.

The Accounttax and Shippingsettings services have functionality for both MCAs
and non-MCAs, but they are distinct, so we just have the main demo entry point
fork depending on the type of client.

Also alter code that used to ignore the return from Update/Patch methods and
call Get immediately afterwards to instead use the return from Update/Patch.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158545679