Open a browser for OAuth2 authorization as needed.
We use the Launchy gem to handle opening a browser to the authorization URL,
and we use the built-in WEBrick HTTP server to receive the redirect with the
refresh token. Now users of the samples don't need to manually copy/paste the
authentication URL and the returned refresh token if using OAuth2 client IDs.

Also, move the authentication code into its own file, and also move token
storage out of the configuration file into a separate file in the
configuration directory. Due to this, the configuration file is now read-only.

Also, we stopped reading websiteUrl from the configuration a while ago, so
make that official in the configuration object.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158708939