Switch from oauth2client to google-auth.
This change also stops using the (possibly non-existent) configuration file as
a backing store for tokens and instead just stores the last successfully
requested one in a separate file.

Since we're guaranteed that the access token is fresh if we get out of the
sample initialization, get rid of all the expired token checks in each of
the method samples and only check for invalid refresh tokens in the Storage
class.

Also change from "native" HTTP batching to using the custombatch methods
where we still use the former.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158975579