Fix null pointer access if no tax rules.
In the JSON response, the API doesn't include fields that would have an
empty array value. This translates to null, not an empty collection, so
we should check for field existence, not the size of its value.

Leave the .Count check in just as a failsafe in case this changes,
though, since both are valid object states.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159720139