Fix comparison between on-wire and from-config values.
Merchant Center IDs coming from the API are represented as strings in
the JSON response. The Merchant Center ID in the JSON configuration
(if used) is represented as a numeric value. Thus, convert the JSON
configuration value to a string so that all MCIDs are appropriately
comparable.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164136432