Add optional logging of responses and requests to Java samples.
Adds a new command line argument, -l/--log_file, which takes a filename and
logs all HTTP requests and responses to that file.

Adds LoggingHttpRequestInitializer, which wraps an existing
HttpRequestInitializer (like the authentication credential) and turns on
logging for any requests and responses that go over the wire.

Extend the standard builder creation to take the parsed arguments (so it can
pass them along to check and see if a logging filename was given) and to also
install a LoggingHttpRequestInitializer around the authentication credentials
if requested.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170529774